refactor: added cytoscape util file to reduce bloat of GraphView
