+++ src/lib/components/GraphView.svelte
@@ -4,100 +4,13 @@
 	import type { GraphData } from '$lib/types/graph';
 	import { selectedNodesStore } from '$lib/stores/selectedNodes';
 	import { appSettings } from '$lib/stores/appSettings';
-	import { formatHex, oklch } from 'culori';
+	import { getLayoutOptions } from '$lib/utils/cytoscape';
+	import { applyCytoscapeStyle } from '$lib/utils/cytoscape';
 
 	export let graphData: GraphData;
 
 	let container: HTMLDivElement;
 	let cy: cytoscape.Core;
-
-	const getResolvedColor = (className: string): string => {
-		const el = document.createElement('div');
-		el.className = `${className} fixed -top-[9999px] w-0 h-0`;
-		document.body.appendChild(el);
-
-		const color = getComputedStyle(el).backgroundColor;
-
-		document.body.removeChild(el);
-		return formatHex(oklch(color)) ?? '#000000';
-	};
-
-	const getLayoutOptions = (settings: typeof $appSettings) => {
-		const baseOptions = {
-			name: settings.layoutName,
-			animate: true,
-			animationDuration: 500,
-			animationEasing: 'ease-in-out'
-		};
-
-		const layoutOptions = settings.layout[settings.layoutName];
-
-		return {
-			...baseOptions,
-			...layoutOptions
-		};
-	};
-
-	function applyCytoscapeStyle(settings: typeof $appSettings) {
-		if (!cy) return;
-		// Resolve colors dynamically based on the current DaisyUI theme
-		const accentColor = getResolvedColor('bg-accent');
-		const neutralColor = getResolvedColor('bg-neutral');
-		const neutralContentColor = getResolvedColor('bg-neutral-content');
-		const baseContentColor = getResolvedColor('bg-base-content');
-		const base300Color = getResolvedColor('bg-base-300');
-
-		cy.style()
-			.selector('node')
-			.style({
-				label: 'data(label)',
-				'background-color': baseContentColor, // node colour
-				color: base300Color,
-				'text-background-color': baseContentColor,
-				'text-background-shape': 'roundrectangle',
-				'text-valign': 'center',
-				'text-halign': 'center',
-				'font-size': settings.nodeFontSize,
-				opacity: settings.nodeOpacity,
-				'border-width': settings.nodeBorderWidth,
-				'text-wrap': settings.wrapText ? 'wrap' : 'none',
-				'text-max-width': settings.wrapText ? `${settings.nodeTextMaxWidth}px` : '0px',
-				'text-background-opacity': 1,
-				'text-background-padding': settings.nodeTextBackgroundPadding
-			})
-			.selector('edge')
-			.style({
-				label: 'data(label)',
-				'curve-style': settings.curveStyle as cytoscape.Css.Edge['curve-style'],
-				'target-arrow-shape': 'triangle',
-				'line-color': neutralColor,
-				'target-arrow-color': neutralColor,
-				color: neutralContentColor, // line colour
-				'text-rotation': 'autorotate',
-				'font-size': settings.edgeFontSize,
-				width: settings.edgeWidth,
-				'text-wrap': 'wrap',
-				'text-max-width': `${settings.edgeTextMaxWidth}px`
-			})
-			.selector('.faded')
-			.style({
-				opacity: 0.1
-			})
-			.selector('node:selected')
-			.style({
-				'border-width': settings.selectedNodeBorderWidth,
-				'border-color': accentColor,
-				'border-opacity': 1,
-				'font-size': settings.selectedNodeFontSize
-			})
-			.selector('edge:selected')
-			.style({
-				'line-color': accentColor,
-				'target-arrow-color': accentColor,
-				width: settings.selectedEdgeWidth
-			})
-			.update();
-	}
 
 	onMount(() => {
 		cy = cytoscape({
@@ -112,7 +25,6 @@
 
 		appSettings.subscribe(async (settings) => {
 			await tick();
-
 			const currentLayoutConfig = settings.layout[settings.layoutName];
 
 			if (
@@ -124,23 +36,19 @@
 				previousLayoutConfig = { ...currentLayoutConfig };
 			}
 
-			applyCytoscapeStyle(settings);
+			applyCytoscapeStyle(cy, settings);
 		});
 
 		cy.on('tap', (event) => {
 			const target = event.target;
-
 			if (target === cy) {
 				cy.elements().removeClass('selected').removeClass('faded');
 				return;
 			}
-
 			if (target.isNode() || target.isEdge()) {
 				selectedNodesStore.addNode({ data: target.data() });
-
 				cy.elements().removeClass('selected').removeClass('faded');
 				target.addClass('selected');
-
 				const connected = target.closedNeighborhood();
 				cy.elements().difference(connected).addClass('faded');
 			}
